Refactor and add 16B alignment check
